Moves the KG builder code to the experimental folder
