+++ docs/source/api.rst
@@ -12,55 +12,55 @@ Components
 KGWriter
 ========
 
-.. autoclass:: neo4j_genai.components.kg_writer.KGWriter
+.. autoclass:: neo4j_genai.experimental.components.kg_writer.KGWriter
     :members: run
 
 Neo4jWriter
 ===========
 
-.. autoclass:: neo4j_genai.components.kg_writer.Neo4jWriter
+.. autoclass:: neo4j_genai.experimental.components.kg_writer.Neo4jWriter
     :members: run
 
 TextSplitter
 ============
 
-.. autoclass:: neo4j_genai.components.text_splitters.base.TextSplitter
+.. autoclass:: neo4j_genai.experimental.components.text_splitters.base.TextSplitter
     :members: run
 
 LangChainTextSplitterAdapter
 ============================
 
-.. autoclass:: neo4j_genai.components.text_splitters.langchain.LangChainTextSplitterAdapter
+.. autoclass:: neo4j_genai.experimental.components.text_splitters.langchain.LangChainTextSplitterAdapter
     :members: run
 
 LlamaIndexTextSplitterAdapter
 =============================
 
-.. autoclass:: neo4j_genai.components.text_splitters.llamaindex.LlamaIndexTextSplitterAdapter
+.. autoclass:: neo4j_genai.experimental.components.text_splitters.llamaindex.LlamaIndexTextSplitterAdapter
     :members: run
 
 TextChunkEmbedder
 =================
 
-.. autoclass:: neo4j_genai.components.embedder.TextChunkEmbedder
+.. autoclass:: neo4j_genai.experimental.components.embedder.TextChunkEmbedder
     :members: run
 
 SchemaBuilder
 =============
 
-.. autoclass:: neo4j_genai.components.schema.SchemaBuilder
+.. autoclass:: neo4j_genai.experimental.components.schema.SchemaBuilder
     :members: run
 
 EntityRelationExtractor
 =======================
 
-.. autoclass:: neo4j_genai.components.entity_relation_extractor.EntityRelationExtractor
+.. autoclass:: neo4j_genai.experimental.components.entity_relation_extractor.EntityRelationExtractor
     :members: run
 
 LLMEntityRelationExtractor
 ==========================
 
-.. autoclass:: neo4j_genai.components.entity_relation_extractor.LLMEntityRelationExtractor
+.. autoclass:: neo4j_genai.experimental.components.entity_relation_extractor.LLMEntityRelationExtractor
     :members: run
 
 .. _retrievers-section:
@@ -219,11 +219,11 @@ Errors
 
   * :class:`neo4j_genai.exceptions.LLMGenerationError`
 
-  * :class:`neo4j_genai.pipeline.exceptions.PipelineDefinitionError`
+  * :class:`neo4j_genai.experimental.pipeline.exceptions.PipelineDefinitionError`
 
-  * :class:`neo4j_genai.pipeline.exceptions.PipelineMissingDependencyError`
+  * :class:`neo4j_genai.experimental.pipeline.exceptions.PipelineMissingDependencyError`
 
-  * :class:`neo4j_genai.pipeline.exceptions.PipelineStatusUpdateError`
+  * :class:`neo4j_genai.experimental.pipeline.exceptions.PipelineStatusUpdateError`
 
 
 Neo4jGenAiError
@@ -327,19 +327,19 @@ LLMGenerationError
 PipelineDefinitionError
 =======================
 
-.. autoclass:: neo4j_genai.pipeline.exceptions.PipelineDefinitionError
+.. autoclass:: neo4j_genai.experimental.pipeline.exceptions.PipelineDefinitionError
    :show-inheritance:
 
 
 PipelineMissingDependencyError
 ==============================
 
-.. autoclass:: neo4j_genai.pipeline.exceptions.PipelineMissingDependencyError
+.. autoclass:: neo4j_genai.experimental.pipeline.exceptions.PipelineMissingDependencyError
    :show-inheritance:
 
 
 PipelineStatusUpdateError
 =========================
 
-.. autoclass:: neo4j_genai.pipeline.exceptions.PipelineStatusUpdateError
+.. autoclass:: neo4j_genai.experimental.pipeline.exceptions.PipelineStatusUpdateError
    :show-inheritance:
