+++ src/neo4j_genai/experimental/components/text_splitters/llamaindex.py
@@ -16,8 +16,8 @@ from __future__ import annotations
 
 from llama_index.core.node_parser import TextSplitter as LlamaIndexTextSplitter
 
-from neo4j_genai.components.text_splitters.base import TextSplitter
-from neo4j_genai.components.types import TextChunk, TextChunks
+from neo4j_genai.experimental.components.text_splitters.base import TextSplitter
+from neo4j_genai.experimental.components.types import TextChunk, TextChunks
 
 
 class LlamaIndexTextSplitterAdapter(TextSplitter):
@@ -32,8 +32,8 @@ class LlamaIndexTextSplitterAdapter(TextSplitter):
     .. code-block:: python
 
         from llama_index.core.node_parser.text.sentence import SentenceSplitter
-        from neo4j_genai.components.text_splitters.langchain import LangChainTextSplitterAdapter
-        from neo4j_genai.pipeline import Pipeline
+        from neo4j_genai.experimental.components.text_splitters.langchain import LangChainTextSplitterAdapter
+        from neo4j_genai.experimental.pipeline import Pipeline
 
         pipeline = Pipeline()
         text_splitter = LlamaIndexTextSplitterAdapter(SentenceSplitter())
