+++ src/neo4j_genai/experimental/pipeline/pipeline_graph.py
