+++ src/neo4j_genai/experimental/pipeline/stores.py
