+++ src/neo4j_genai/experimental/pipeline/types.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 
 from pydantic import BaseModel, ConfigDict
 
-from neo4j_genai.pipeline.component import Component
+from neo4j_genai.experimental.pipeline.component import Component
 
 
 class ComponentConfig(BaseModel):
