+++ tests/e2e/test_kg_construction_e2e.py
@@ -15,8 +15,8 @@
 
 import neo4j
 import pytest
-from neo4j_genai.components.kg_writer import Neo4jWriter
-from neo4j_genai.components.types import (
+from neo4j_genai.experimental.components.kg_writer import Neo4jWriter
+from neo4j_genai.experimental.components.types import (
     Neo4jGraph,
     Neo4jNode,
     Neo4jRelationship,
