+++ tests/unit/experimental/components/test_schema.py
@@ -15,13 +15,13 @@
 from __future__ import annotations
 
 import pytest
-from neo4j_genai.components.schema import (
+from neo4j_genai.exceptions import SchemaValidationError
+from neo4j_genai.experimental.components.schema import (
     SchemaBuilder,
     SchemaEntity,
     SchemaProperty,
     SchemaRelation,
 )
-from neo4j_genai.exceptions import SchemaValidationError
 from pydantic import ValidationError
