+++ tests/unit/experimental/components/text_splitters/test_langchain.py
@@ -14,8 +14,10 @@
 #  limitations under the License.
 import pytest
 from langchain_text_splitters import RecursiveCharacterTextSplitter
-from neo4j_genai.components.text_splitters.langchain import LangChainTextSplitterAdapter
-from neo4j_genai.components.types import TextChunk, TextChunks
+from neo4j_genai.experimental.components.text_splitters.langchain import (
+    LangChainTextSplitterAdapter,
+)
+from neo4j_genai.experimental.components.types import TextChunk, TextChunks
 
 text = """
 Lorem ipsum dolor sit amet, consectetur adipiscing elit.
