+++ tests/unit/experimental/pipeline/test_pipeline_graph.py
@@ -13,7 +13,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 import pytest
-from neo4j_genai.pipeline.pipeline_graph import (
+from neo4j_genai.experimental.pipeline.pipeline_graph import (
     PipelineEdge,
     PipelineGraph,
     PipelineNode,
